Restore automated deploy
We fixed some version about node and nx. After this we will need to
migrate workspace due to several major version of nx (right now the
projects runs on nx 10 instead of nx 13).
Also we removed fake test steps (no tests were actually executed). We
will try to restore them once we fixed them (most of them are outdated).

We now build js files using the same container we create to serve debug
and run the application, so we are sure to use the very same nodejs
version.