Remove popup test
This test will never work. To make it work we should
refactor the app to avoid rendering popup outside the
component itself. We tried for a couple of hours but hit
a series of side effects. At the moment of writing it does
not make sense to invest more time on it.